requirements.txt
# Flask Framework
Flask==2.0.1
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.32
SQLAlchemy-Utils==0.41.2

# Configuration
python-dotenv==1.0.1

# Redis for caching and sessions
redis==4.6.0

# RabbitMQ for message queuing
pika==1.3.2

# Testing
pytest==6.2.4
blinker==1.4
exceptiongroup==1.0.0rc9
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
tomli==1.2.1

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2